Initialise repository with rescued code
I managed to rescue this code after my laptop was unintentionally wiped
(don’t ask about it). I had a git repository, but stupidly forgot to
uploaded it to GitHub or put it in my Google Drive folder. So now all
that history is gone -- the code is still here, luckily.